backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.3
google-generativeai==0.3.2
openai==1.10.0
elevenlabs==0.2.26
httpx==0.26.0
